changed int to \float

+++ General-Function-Codes/phone_data.py
@@ -15,7 +15,7 @@ today = datetime.date.today()
 # %%
 
 data_today = input("Enter data remaining (GB) :  ")
-data_today = int(data_today)
+data_today = float(data_today)
 #data_today = 38
 days_remain = end_date - today
 data_to_remain = 2
